Extracting common code for rewards tests
